Large Commit
Havent made a commit in a long time, partially due to lack of a strong
internet connection. This commit has the updated code allowing for the
camera to work as well as corrections for stability issues, image
compression was also introduced to lighten the load on the users mobile
data.